Cleanup /extract/finish share login endpoint
