+++ api/src/auth/drivers/local.ts
@@ -1,6 +1,5 @@
 import { Router } from 'express';
 import argon2 from 'argon2';
-import ms from 'ms';
 import Joi from 'joi';
 import { AuthDriver } from '../auth';
 import { User } from '../../types';
@@ -9,7 +8,7 @@ import { AuthenticationService } from '../../services';
 import asyncHandler from '../../utils/async-handler';
 import env from '../../env';
 import { respond } from '../../middleware/respond';
-import { DIRECTUS_SHARED_AUTH } from '../../constants';
+import { COOKIE_OPTIONS } from '../../constants';
 
 export class LocalAuthDriver extends AuthDriver {
 	async getUserID(payload: Record<string, any>): Promise<string> {
@@ -51,14 +50,6 @@ export function createLocalAuthRouter(provider: string): Router {
 		otp: Joi.string(),
 	}).unknown();
 
-	const sharedLoginSchema = Joi.object({
-		mode: Joi.string().valid('cookie', 'json'),
-		id: Joi.string().required(),
-		password: Joi.string(),
-	}).unknown();
-
-	const loginSchema = provider === DIRECTUS_SHARED_AUTH ? sharedLoginSchema : userLoginSchema;
-
 	router.post(
 		'/',
 		asyncHandler(async (req, res, next) => {
@@ -73,7 +64,7 @@ export function createLocalAuthRouter(provider: string): Router {
 				schema: req.schema,
 			});
 
-			const { error } = loginSchema.validate(req.body);
+			const { error } = userLoginSchema.validate(req.body);
 
 			if (error) {
 				throw new InvalidPayloadException(error.message);
@@ -96,13 +87,7 @@ export function createLocalAuthRouter(provider: string): Router {
 			}
 
 			if (mode === 'cookie') {
-				res.cookie(env.REFRESH_TOKEN_COOKIE_NAME, refreshToken, {
-					httpOnly: true,
-					domain: env.REFRESH_TOKEN_COOKIE_DOMAIN,
-					maxAge: ms(env.REFRESH_TOKEN_TTL as string),
-					secure: env.REFRESH_TOKEN_COOKIE_SECURE ?? false,
-					sameSite: (env.REFRESH_TOKEN_COOKIE_SAME_SITE as 'lax' | 'strict' | 'none') || 'strict',
-				});
+				res.cookie(env.REFRESH_TOKEN_COOKIE_NAME, refreshToken, COOKIE_OPTIONS);
 			}
 
 			res.locals.payload = payload;
